Move get_member
