+++ src/commands/tags/_tags_create.ts
@@ -2,8 +2,8 @@ import { Message } from 'discord.js';
 import { TAG_CREATE_PERMITTED_IDS } from '../../config.js';
 import { supabase } from '../../db/index.js';
 import { tags_embed_builder } from '../../utils/embed_helpers.js';
-import { has_any_role_or_id } from '../../utils/snowflake.js';
-import { get_member, get_tag, Tag, TagCUDHandler } from './_common.js';
+import { get_member, has_any_role_or_id } from '../../utils/snowflake.js';
+import { get_tag, Tag, TagCUDHandler } from './_common.js';
 
 const validator_regex = /^[a-z0-9\-\+\_\.\ ]*$/;
