+++ src/utils/snowflake.ts
@@ -1,4 +1,4 @@
-import { GuildMember, Snowflake } from 'discord.js';
+import { CommandInteraction, GuildMember, Snowflake } from 'discord.js';
 
 /**
  * Check if a member has any of the given roles or ids.
@@ -16,4 +16,14 @@ export function has_any_role_or_id(
 	}
 
 	return false;
+}
+
+/**
+ * Get the member having the `id` in the guild where the interaction was created.
+ */
+export async function get_member(
+	interaction: CommandInteraction,
+	id: Snowflake,
+) {
+	return interaction.guild?.members.fetch(id);
 }
